new structure

+++ scripts/process_submissions.py
@@ -3,11 +3,15 @@ import csv
 import shutil
 import subprocess
 import magic
-import pandas as pd
 from convert_to_html import convert_to_html
 
+BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+DATA_DIR = os.path.join(BASE_DIR, '..', 'data')
+SUBMISSIONS_DIR = os.path.join(BASE_DIR, '..', 'submissions')
+ARCHIVES_DIR = os.path.join(BASE_DIR, '..', 'archives')
 
-def csv_create(base_dir="submissions/", csv_file="students_submissions.csv"):
+
+def csv_create(base_dir=SUBMISSIONS_DIR, csv_file=os.path.join(DATA_DIR, "students_submissions.csv")):
     students_info = []
 
     for folder_name in os.listdir(base_dir):
@@ -19,7 +23,7 @@ def csv_create(base_dir="submissions/", csv_file="students_submissions.csv"):
                 [student_name, unique_id, "", "", "", "", "", ""])
 
     headers = ["Student Name", "Unique Identifier", "Folder Structure",
-               "Compiled", "Compile Log", "Error Log", "Changed Files", "Deleted files"]
+               "Compiled", "Compile Log", "Error Log", "Changed Files", "Deleted Files"]
 
     with open(csv_file, mode='w', newline='', encoding='utf-8-sig') as file:
         writer = csv.writer(file)
@@ -54,13 +58,11 @@ def extract_archive(file_path, destination_dir):
 
         if not extracted:
             if file_path.endswith('.rar'):
-                # Try unzipping if the extension is .rar but it might be a zip file
                 new_file_path = file_path.replace('.rar', '.zip')
                 os.rename(file_path, new_file_path)
                 shutil.unpack_archive(new_file_path, destination_dir)
                 extracted = True
             elif file_path.endswith('.zip'):
-                # Try unzipping if the extension is .zip but it might be a rar file
                 new_file_path = file_path.replace('.zip', '.rar')
                 os.rename(file_path, new_file_path)
                 os.makedirs(destination_dir, exist_ok=True)
@@ -90,12 +92,27 @@ def clean_up_folder(folder_path):
             file_path = os.path.join(root, file)
             if file == "input.txt":
                 continue
+            new_file_path = None
             if not is_source_code(file_path):
                 try:
-                    os.remove(file_path)
-                    deleted_files_log.append(file)
+                    if file.endswith('.txt') and 'c' in file_path:
+                        new_file_path = file_path.replace('.txt', '.c')
+                    elif file.endswith('.txt') and 'cpp' in file_path:
+                        new_file_path = file_path.replace('.txt', '.cpp')
+                    elif file.endswith('.rar'):
+                        new_file_path = file_path.replace('.rar', '.zip')
+                    elif file.endswith('.zip'):
+                        new_file_path = file_path.replace('.zip', '.rar')
+
+                    if new_file_path:
+                        os.rename(file_path, new_file_path)
+                        changed_files_log.append(
+                            f"{file} -> {os.path.basename(new_file_path)}")
+                    else:
+                        os.remove(file_path)
+                        deleted_files_log.append(file)
                 except Exception as e:
-                    print(f"Error deleting file {file_path}: {e}")
+                    print(f"Error processing file {file_path}: {e}")
     return deleted_files_log, changed_files_log
 
 
@@ -109,25 +126,29 @@ def find_all_source_files(student_folder):
     return source_files
 
 
-def compile_check(source_files, student_folder):
+def compile_check(source_files):
     compile_results = {}
     compile_logs = {}
+    error_logs = {}
     for file_path in source_files:
         if os.path.isfile(file_path):  # Ensure it's a file
-            compile_command = f"gcc \"{file_path}\" -o \"{file_path}.out\""
+            if file_path.endswith('.cpp'):
+                compile_command = f"g++ \"{file_path}\" -o \"{file_path}.out\""
+            else:
+                compile_command = f"gcc \"{file_path}\" -o \"{file_path}.out\""
             try:
                 result = subprocess.run(
                     compile_command, shell=True, capture_output=True, text=True)
                 if result.returncode == 0:
                     compile_results[file_path] = True
-                    compile_logs[file_path] = "Compiled successfully."
+                    compile_logs[file_path] = result.stdout
                 else:
                     compile_results[file_path] = False
-                    compile_logs[file_path] = result.stderr
+                    error_logs[file_path] = result.stderr
             except Exception as e:
                 compile_results[file_path] = False
-                compile_logs[file_path] = str(e)
-    return compile_results, compile_logs
+                error_logs[file_path] = str(e)
+    return compile_results, compile_logs, error_logs
 
 
 def ensure_input_txt(student_folder, root_input_txt):
@@ -148,12 +169,12 @@ def log_folder_structure(folder_path):
     return "\n".join(folder_structure)
 
 
-def update_csv_with_errors(base_dir="submissions/", csv_file="students_submissions.csv"):
+def update_csv_with_errors(base_dir=SUBMISSIONS_DIR, csv_file=os.path.join(DATA_DIR, "students_submissions.csv"), check_input_txt=True):
     with open(csv_file, mode='r', newline='', encoding='utf-8-sig') as file:
         reader = csv.reader(file)
         rows = list(reader)
 
-    root_input_txt = "input.txt"
+    root_input_txt = os.path.join(DATA_DIR, "input.txt")
 
     for row in rows[1:]:
         student_folder = os.path.join(
@@ -173,7 +194,9 @@ def update_csv_with_errors(base_dir="submissions/", csv_file="students_submissio
                 initial_structure = log_folder_structure(student_folder)
                 row[2] = initial_structure
 
-                ensure_input_txt(student_folder, root_input_txt)
+                if check_input_txt:
+                    ensure_input_txt(student_folder, root_input_txt)
+
                 deleted_files_log, changed_files_log = clean_up_folder(
                     student_folder)
                 row[7] = ", ".join(deleted_files_log)
@@ -181,15 +204,18 @@ def update_csv_with_errors(base_dir="submissions/", csv_file="students_submissio
 
                 source_files = find_all_source_files(student_folder)
                 if source_files:
-                    compile_results, compile_logs = compile_check(
-                        source_files, student_folder)
+                    compile_results, compile_logs, error_logs = compile_check(
+                        source_files)
                     row[3] = ", ".join(
                         [f"{k}: {v}" for k, v in compile_results.items()])
                     row[4] = "\n".join(
                         [f"{k}: {v}" for k, v in compile_logs.items()])
+                    row[5] = "\n".join(
+                        [f"{k}: {v}" for k, v in error_logs.items()])
                 else:
                     row[3] = "No source files found"
                     row[4] = "N/A"
+                    row[5] = ""
 
             except Exception as e:
                 row[5] += str(e)
@@ -199,13 +225,13 @@ def update_csv_with_errors(base_dir="submissions/", csv_file="students_submissio
         writer.writerows(rows)
 
 
-def first_control():
+def first_control(check_input_txt=True):
     csv_create()
-    update_csv_with_errors()
+    update_csv_with_errors(check_input_txt=check_input_txt)
     convert_to_html()
     print("First control finished.")
 
 
 if __name__ == "__main__":
-    first_control()
+    first_control(check_input_txt=True)
